Added separate FQ decomp pass
